requirements.txt
paho-mqtt==1.5.0
pymavlink==2.4.8
dronekit==2.9.2
pyserial==3.4
requests==2.23.0
mavproxy==1.8.19
huawei-lte-api==1.4.12
